Muster roll type error fixes

+++ frontend/works_shg_app/lib/models/muster_rolls/muster_roll_model.dart
@@ -46,7 +46,7 @@ class IndividualEntries with _$IndividualEntries {
   const factory IndividualEntries({
     String? id,
     String? individualId,
-    int? totalAttendance,
+    double? totalAttendance,
     @JsonKey(name: 'attendanceEntries')
         List<AttendanceEntries>? attendanceEntries,
   }) = _IndividualEntries;
